requirements.txt
psycopg2
pdfminer.six
environ
lxml


